docs: brief comments on implementation
