+++ packages/motif/src/components/Legend/type.ts
@@ -2,11 +2,13 @@ import { MouseEventHandler } from 'react';
 
 export type ColorMaps = [string, string];
 export type LegendProps = {
-  data: { [_: string]: string };
+  data: Record<string, string>;
   colorMap: string[];
   kind: 'node' | 'edge';
   maxSize?: number;
   label?: string;
+
+  // perform attributes colour changes
   onChangeColor?: (target: ColorMaps) => void;
   isAllowChangeColor?: boolean;
 };
@@ -19,3 +21,17 @@ export type GraphAttributeColourProps = {
 
 export type ObjectColourProps = GraphAttributeColourProps &
   Pick<LegendProps, 'kind'> & { label: string };
+
+export type ObjectColourPickerProps = {
+  // selected attribute (node/edge) to modify the color when user clicked on the element.
+  selectedAttr: ColorMaps;
+
+  // function provided to change the node/edge color
+  onChangeColor: (target: ColorMaps) => void;
+
+  // confirm the current colour changes.
+  onComplete: (confirmedColor: ColorMaps) => void;
+
+  // revert the default colour when user decided to discard the changes.
+  onCancel: (defaultColor: ColorMaps) => void;
+};
